refactor(dashboard): split monolithic component and optimize layout
- Split 1077-line Dashboard into 6 focused sub-components:
  OverviewCards, ResourceDistribution, OperationAnalysis,
  SystemStatus, TrendCharts, RealtimeTables
- Separate statistics into two independent cards:
  "Resource Distribution" (users + devices) and
  "Operation Analysis" (connections + files)
- Extract SystemStatus as standalone card with Progress bars
  instead of cramped Tags in trend chart header
- Increase trend chart height from 220px to 280px
- Simplify overview cards to show 1 key sub-indicator each
- Add i18n keys for new section titles (resourceDistribution,
  operationAnalysis) in both en-US and zh-CN
